factor out Database
